Create components for card items

+++ app/src/main/java/com/example/cfttesttask/data/local/dbo/CountryDbo.kt
@@ -6,8 +6,8 @@ data class CountryDbo(
     val alpha2: String?,
     val currency: String?,
     val emoji: String?,
-    val latitude: Int?,
-    val longitude: Int?,
+    val latitude: Double?,
+    val longitude: Double?,
 
     @ColumnInfo(name = "countryName")
     val name: String?,
